src/pysnip/__init__.py
__version__ = "7.3"
__date__ = ""

from zuper_commons.logs import ZLogger, ZLoggerInterface

logger: ZLoggerInterface = ZLogger(__name__)
logger.hello_module(name=__name__, filename=__file__, version=__version__, date=__date__)

from .write_source import *
from .job import *
from .main import *
from .meat import *
from .capture import *
from .lenient_option_parser import *
from .script_utils import *

logger.hello_module_finished(__name__)
