pyproject.toml
[project]
name = "PySnip-z7"
version = "7.2"
authors = [{name = "Andrea Censi", email = "AndreaCensi@users.noreply.github.com"}]
dependencies = [
    "compmake-z7",
    "zuper-commons-z7",
    "zuper-utils-asyncio-z7",
    "zuper-zapp-z7",
    ]

[project.scripts]
pysnip-make = "pysnip:pysnip_make_main"

[tool]
[tool.poetry]
name = "PySnip-z7"
version = "7.2"
description = ""
authors = ["Andrea Censi <AndreaCensi@users.noreply.github.com>"]
packages = [
    {from = "src", include = "pysnip"},
    ]

[tool.setuptools]
packages = [
    "pysnip",
    ]
package-dir = {"" = "src"}
